Drop Authen::SASL dependency
Under certain circumstances it could cause warnings/errors related to
Authen::SASL::XS, which does not like the way Cassandra handles
authentication. Upon further inspection, it turns out that Cassandra's
authentication is not very SASL-ish and its implementation violates the
SASL specification in several ways.

With this patch we now just mimic the Java driver's implementation,
which is to just concatenate the username and password into a
null-delimited string.