Cassandra::Client 0.04 + DBD::Cassandra 0.51
I've been developing these for a while and went through several
iterations to get these working. This is a *full rewrite* of the entire
driver, but DBD::Cassandra retains most unit test compatibility, with
the exception of being able to do "use <keyspace>" in a query.

Splitting Cassandra::Client from DBD::Cassandra allows giving the client
a more Cassandra-native interface. The actual interface was inspired by
the node.js interface.

This also introduces very basic multi-node support, to be extended in
future versions. Right now we just create a pool of X (default 2)
servers, and round-robin when doing queries.

In my tests this has given a very significant speedup in asynchronous
queries, without changing any of the code using the API.

DBD::Cassandra is now merely a wrapper around the real client,
Cassandra::Client. It's up to the users to pick which one they prefer,
I'll maintain both (which is now easy).

Enjoy!

PS: The driver is unfinished, but can be used safely in production. For
example, "wait for schema change" is implemented as "sleep 1", but the
previous code had no such code anyway :-)