Initial client throttling implementation
First attempt at writing something to stop crazy clients from causing
catastrophic failures on clusters. The algorithm should be sound, but
the places where I'm hooking into the logic are probably not.

More code for this will follow later :-) This is an undocumented feature
for now, so I can do some experimentation on it.