Turn utf8 flag off for blobs and other "passthru" style inputs
Because of how Perl deals with utf8 flags, once a flagged input is used
for any of the passthru types (blob, ascii, custom), it could go into
the "pending write buffer" and infect other data. This would then cause
completely random "Request timed out" errors.

DBD-Cassandra/t/30-types.t found this one, but I had to run it a lot of
times for the issue to show.