Initial work for command queueing
Async APIs are fun, but at some point you want to slow things down a
bit. Right now we have the Throttler code, to ensure we react to
timeouts, we have the Retry code, to handle them, and the Queue code, to
ensure we only do X concurrent queries.

That's the last one of the major features I wanted to implement. The
three systems need some work, but now they actually exist. :-)