Overhaul how we mark retryable requests, and timeouts
I discovered that AdaptiveThrottler was counting the wrong errors, so
that's now fixed. I also made sure we retry only the things we know we
can safely retry, instead of just randomly retrying all the things. Oh,
and I special-cased the "we failed to send the request" case, because
that seems to be the most likely place where we notice situations like
rolling restarts.