Cassandra::Client: perl 5.20+ support, I hope
Tested with 5.24 and it seems to be happy. An optimization of mine
backfired because Perl decided in 5.20 to optimize something else, and
the two clashed.

I've added a unit test for the async utils, which will work even without
a server (unlike most tests :/) and used that test as a benchmark to
figure out the best way to work around the 5.20 issue.