Date support: because it's cool to write your own date library
(No it's not)

Always wanted to do this. Passes the unit tests. Unit tests are
probably broken. ;-)

The actual implementation is O(1) and based on Julian Day calculations,
a concept very similar to how Cassandra stores dates internally.

Btw, I tried testing for edge cases in the far past and far future, but
the Python client doesn't seem to implement anything too far in the
past or future.