Handle UTF-8 better during authentication
In the current situation UTF-8 auth may work, but will trigger warnings
and potentially corrupt the internal state if it's done in a certain
way. That's not ideal. Now I don't know why anyone would use UTF-8 auth,
but let's support it.

This also adds some sanity checks to make sure we don't run into this in
the future either.