Try harder to satisfy requests when we get disconnected unexpectedly
During a rolling restart you'd want to remain connected to the cluster
but not the host. In theory having a single host down at a time should
mean that with our default of max_connections=2 we should remain
connected -- but as it turns out this is not the case when we don't do
enough requests to Cassandra to actually notice that something happened
(because our event loop may only be invoked when we do queries).

By saying "try twice" instead of "try once" we can cover this case a bit
better.