Allow more reconnection attempts
This one is getting rather interesting... When using the EV eventloop
(though AnyEvent is also affected) it's possible that the callbacks are
only invoked something like once a minute, for example when there just
aren't a lot of queries to execute (idle webservers, etc).

During a rolling restart it's possible that all connections get
disconnected, and when finally doing a query again we first need to
connect. But if we already had a pending reconnection the last time we
executed a query and simply haven't gotten back to it, our timeouts can
kick in and immediately consider all those reconnection attempts failed.
We then count these timeouts towards the budget for connection attempts
and immediately fail the queries for timeout reasons, even if they've
been executed a ms ago.

I'm not 100% sure, but this issue should affect the EV (= default)
backend only when using more than two concurrent connections, but it can
affect the AnyEvent backend when multiple Cassandra::Client objects are
in use and both are only rarely queried -- or when other AnyEvent
modules are being used.

Of course the real fix should be to not consider those connection
attempts a timeout -- but I'm first going for this solution which we'll
need anyway.