Always clean our timeout list when disconnecting
This is a bit nasty... Although we do properly disarm all the deadlines, they're
still in the list of timeouts and therefore we don't shutdown our AE/EV timers.
So when disconnecting in the middle of a networkstatus update (or other type of
active query), we'll accidentally keep the timer alive, which drags AE/EV along
with it, resulting in a memory leak, and worse, a fd leak (when AE/EV uses
eg. epoll).