Fix where response for async query gets overwritten by another's
Using goto caused the $body variable to become dynamically scoped, which
meant that a callback might not have had time to deal with it before it
got overwritten when decoding the next response.

On high-concurrent requests, that manifested as the response for a
request being replaced with a response for a different request.

I.e., an async_execute() response could be for a different
async_execute().

Other possible fixes are:
1. Moving the $body parsing to a different method;
2. Creating a copy of $body before calling the $cb.

This fix was chosen as I don't know if other users of $body could have
deferred execution (so 2. might not be enough); or if there would be
performance concerns with 1.